pyproject.toml
[project]
authors = [{ name = "Ahmad Akra", email = "ahakra@outlook.com" }]
dependencies = [
    "transformers (>=4.51.3,<5.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "bson (>=0.5.10,<0.6.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "nats-py (>=2.10.0,<3.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
]
description = ""
name = "ai-project-text-analysis"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]


[tool.poetry]
name = "ai-project-text-analysis"
version = "0.1.0"
description = ""
authors = ["ahakra@outlook.com"]
package-mode = false
